Modify Header components, add globalStyles.js
