+++ .prettierrc
@@ -1,4 +1,4 @@
 {
   "singleQuote": true,
-  "endOfLine":"auto"
+  "endOfLine": "auto"
 }
