--- src/components/Header.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import Logo from './Logo';
-import Navigation from './Navigation';
-
-const Header = () => (
-  <>
-    <Logo />
-    <Navigation />
-  </>
-);
-
-export default Header;
