+++ src/components/Header/Header.styles.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+export const Wrapper = styled.header`
+  display: flex;
+  height: 64px;
+  weight: 100vw;
+  background-color: blue;
+`;
