+++ src/components/Navigation/Navigation.styles.js
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+
+export const StyledNavigation = styled.div`
+  display: flex;
+  background-color: green;
+`;
