--- src/components/pages/Root.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import Header from '../Header';
-
-const Root = () => (
-  <div className="wrapper">
-    <Header />
-  </div>
-);
-
-export default Root;
