+++ src/components/views/Root.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import Header from '../Header/Header';
+import GlobalStyle from '../../assets/styles/globalStyles';
+
+const Root = () => (
+  <div>
+    <GlobalStyle />
+    <Header />
+  </div>
+);
+
+export default Root;
